study/Y2026/02_Feb/02th/homework(nara)/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "axios>=0.4.0",
    "fastapi[standard]>=0.128.0",
    "python-jose>=3.5.0",
]
